Add support sql NULL type for User model

+++ usecases/interface.go
@@ -18,7 +18,7 @@ type UserRepository interface {
 
 type User struct {
 	ID        int64  `json:"id"`
-	Age       int    `json:"age"`
+	Age       int64  `json:"age"`
 	FirstName string `json:"first_name"`
 	LastName  string `json:"last_name"`
 	Email     string `json:"email"`
